requirements.txt
discord==1.0.1
requests==2.25.1
python-dotenv==0.17.0
